// apps/web/src/lib/api/generated/model/ratingResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews
 * OpenAPI spec version: 1.0.0
 */
import type { RatingResponsePredefinedDescription } from './ratingResponsePredefinedDescription';

export interface RatingResponse {
  id?: number;
  vetId?: number;
  score?: number;
  description?: string;
  predefinedDescription?: RatingResponsePredefinedDescription;
  customerName?: string;
  rateDate?: string;
}